Add files via upload
Removed leg from bird. This ease 3d printing, and increase brick stability. Including the leg is closer to representation in book, so therefore we may still want the leg.
I currently don't know the rules of the game, but if the player is moving the bird and we keep the leg, we should be careful to strengthen the leg.

Also I suggest removing sharp edges, as the game is for kids.